+++ src/main/java/racingcar/validator/ConsoleInputValidator.java
@@ -1,4 +1,4 @@
-package racingcar.view;
+package racingcar.validator;
 
 import java.util.List;
